feat: add zod validator
