+++ package.json
@@ -11,11 +11,13 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@hono/zod-validator": "^0.7.4",
     "@spotify/web-api-ts-sdk": "^1.2.0",
     "dayjs": "^1.11.18",
     "discord-interactions": "^4.4.0",
     "hono": "4.10.2",
-    "openai": "^6.6.0"
+    "openai": "^6.6.0",
+    "zod": "^4.1.12"
   },
   "devDependencies": {
     "@cloudflare/vitest-pool-workers": "^0.9.14",
