+++ src/routes/index.ts
@@ -1,18 +1,31 @@
 import { Hono } from "hono";
 
-import aiRoute from "./ai";
-
 import type { Bindings } from "~/types/bindings";
 
-import discordRoute from "~/routes/discord";
-import labRoute from "~/routes/lab";
-import spotifyRoute from "~/routes/spotify";
+import generateThemeRoute from "~/routes/ai/generate-theme";
+import commandsRoute from "~/routes/discord/commands";
+import interactionsRoute from "~/routes/discord/interactions";
+import messagesRoute from "~/routes/discord/messages";
+import registerRoute from "~/routes/discord/register";
+import iframeRoute from "~/routes/lab/iframe";
+import postRoute from "~/routes/lab/post";
+import myTopTracksRoute from "~/routes/spotify/my-top-tracks";
+import searchRoute from "~/routes/spotify/search";
 
 const app = new Hono<{ Bindings: Bindings }>()
   .get("/", (c) => c.text("🔥"))
-  .route("/spotify", spotifyRoute)
-  .route("/discord", discordRoute)
-  .route("/ai", aiRoute)
-  .route("/lab", labRoute);
+  // Spotify routes
+  .route("/spotify/my-top-tracks", myTopTracksRoute)
+  .route("/spotify/search", searchRoute)
+  // Discord routes
+  .route("/discord", interactionsRoute)
+  .route("/discord/register", registerRoute)
+  .route("/discord/commands", commandsRoute)
+  .route("/discord/channels", messagesRoute)
+  // AI routes
+  .route("/ai/generate-theme", generateThemeRoute)
+  // Lab routes
+  .route("/lab/post", postRoute)
+  .route("/lab/iframe", iframeRoute);
 
 export default app;
