+++ src/routes/lab/iframe.tsx
@@ -1,5 +1,10 @@
-import type { Context } from "hono";
+import { Hono } from "hono";
 import type { FC } from "hono/jsx";
+
+import { zValidator } from "@hono/zod-validator";
+import { z } from "zod";
+
+import type { Bindings } from "~/types/bindings";
 
 const IFrameComponent: FC<{ url: string }> = (props) => {
   const title = "iframe demo";
@@ -17,10 +22,18 @@ const IFrameComponent: FC<{ url: string }> = (props) => {
   );
 };
 
-const iframeRoute = (c: Context) => {
-  const param = c.req.param("param");
-  const url = URL.canParse(param) ? param : "https://newt239.dev/";
-  return c.html(<IFrameComponent url={url} />);
-};
+const iframeParamSchema = z.object({
+  url: z.string().min(1, "URL is required"),
+});
 
-export default iframeRoute;
+const app = new Hono<{ Bindings: Bindings }>().get(
+  "/:url",
+  zValidator("param", iframeParamSchema),
+  (c) => {
+    const { url: param } = c.req.valid("param");
+    const url = URL.canParse(param) ? param : "https://newt239.dev/";
+    return c.html(<IFrameComponent url={url} />);
+  }
+);
+
+export default app;
