+++ src/routes/lab/post.ts
@@ -1,15 +1,26 @@
-import type { Context } from "hono";
+import { Hono } from "hono";
 
-const postRoute = async (c: Context) => {
-  const body = await c.req.json();
-  const header = {
-    userAgent: c.req.header("User-Agent"),
-    ContentType: c.req.header("Content-Type"),
-  };
-  const res = { body, header };
-  console.log(res);
+import { zValidator } from "@hono/zod-validator";
+import { z } from "zod";
 
-  return c.json(res);
-};
+import type { Bindings } from "~/types/bindings";
 
-export default postRoute;
+const postBodySchema = z.record(z.string(), z.unknown());
+
+const app = new Hono<{ Bindings: Bindings }>().post(
+  "/",
+  zValidator("json", postBodySchema),
+  async (c) => {
+    const body = c.req.valid("json");
+    const header = {
+      userAgent: c.req.header("User-Agent"),
+      ContentType: c.req.header("Content-Type"),
+    };
+    const res = { body, header };
+    console.log(res);
+
+    return c.json(res);
+  }
+);
+
+export default app;
