+++ src/routes/spotify/my-top-tracks.ts
@@ -1,0 +1,30 @@
+import { Hono } from "hono";
+
+import type { Bindings } from "~/types/bindings";
+
+import { createSpotifyClient } from "~/clients/spotify";
+
+const app = new Hono<{ Bindings: Bindings }>().get("/", async (c) => {
+  const client = createSpotifyClient(
+    c.env.SPOTIFY_CLIENT_ID,
+    c.env.SPOTIFY_CLIENT_SECRET
+  );
+  const { items } = await client.currentUser.topItems("tracks", "short_term");
+  if (!items) {
+    return c.json([]);
+  }
+  const returnData = items.map((track) => {
+    return {
+      name: track.name,
+      artists: track.artists.map((artist) => artist.name),
+      thumbnail: track.album.images[0].url,
+      preview: track.preview_url ? track.preview_url : null,
+      duration: track.duration_ms,
+      popularity: track.popularity,
+      link: track.external_urls.spotify,
+    };
+  });
+  return c.json(returnData);
+});
+
+export default app;
